version 2

+++ web/recorder/models.py
@@ -1,5 +1,4 @@
 import os
-from enum import Enum
 from logging import getLogger
 
 from django.conf import settings
@@ -8,16 +7,15 @@ from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.postgres.fields import JSONField
 from django.core.files.base import ContentFile
+from django.core.validators import FileExtensionValidator, MinLengthValidator, URLValidator
 from django.db import models
 from django.utils import timezone
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 
-from ffmpeg.generator import Command
+from command.models import Queue
 from ffmpeg.utils.base import ChoiceEnum
-from ffmpeg.utils.codecs import Codec
 from ffmpeg.utils.ffprobe import get_file_attributes
-from ffmpeg.utils.filters import BitstreamChannelFilter, StreamSpecifier, ScaleFilter, FFmpegFilter
-from ffmpeg.utils.log import LogLevel
+from ffmpeg.utils.filters import FOAR
 from ffmpeg.utils.video import VIDEO_SIZES
 
 User = getattr(settings, 'AUTH_USER_MODEL', get_user_model())
@@ -25,7 +23,8 @@ logger = getLogger('recorder.models')
 
 
 class Category(models.Model):
-    name = models.CharField(verbose_name=_('Category Name'), unique=True, max_length=100)
+    name = models.CharField(verbose_name=_('Category Name'), unique=True, max_length=100,
+                            validators=[MinLengthValidator(2)])
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
@@ -39,11 +38,17 @@ class Category(models.Model):
     class Meta:
         verbose_name = _("Category")
         verbose_name_plural = _("Categories")
+        app_label = "recorder"
+
+    def save(self, **kwargs):
+        self.full_clean()
+        return super(Category, self).save(**kwargs)
 
 
 class Channel(models.Model):
-    name = models.CharField(verbose_name=_('Channel Name'), unique=True, max_length=100)
-    url = models.URLField(verbose_name=_('URL'))
+    name = models.CharField(verbose_name=_('Channel Name'), unique=True, max_length=100,
+                            validators=[MinLengthValidator(2)])
+    url = models.URLField(verbose_name=_('URL'), validators=[URLValidator])
     category = models.ForeignKey('Category', null=True, blank=True, verbose_name=_('Category'))
 
     created_at = models.DateTimeField(auto_now_add=True)
@@ -55,6 +60,10 @@ class Channel(models.Model):
     class Meta:
         verbose_name = _("Channel")
         verbose_name_plural = _("Channels")
+
+    def save(self, **kwargs):
+        self.full_clean()
+        return super(Channel, self).save(**kwargs)
 
 
 class ScheduleStatus(ChoiceEnum):
@@ -68,15 +77,21 @@ class ScheduleStatus(ChoiceEnum):
 
 class Schedule(models.Model):
     channel = models.ForeignKey('Channel', verbose_name=_('Channel'))
-    name = models.CharField(verbose_name=_('Record Name'), max_length=100)
+    name = models.CharField(verbose_name=_('Record Name'), max_length=100, validators=[MinLengthValidator(2)])
     start_time = models.DateTimeField(verbose_name=_('Start Time'))
     time = models.TimeField(verbose_name=_('Time'))
 
     status = models.SmallIntegerField(verbose_name=_('Status'), choices=[(s.value, _(s.name)) for s in ScheduleStatus],
-                                      default=ScheduleStatus.Scheduled.value)
+                                      default=int(ScheduleStatus.Scheduled))
 
-    file = models.ForeignKey('Video', null=True, blank=True)
+    file = models.FileField(verbose_name=_("Video"), null=True, blank=True)
 
+    # Resize Task
+    resize = models.CharField(max_length=10, choices=VIDEO_SIZES, null=True, blank=True, verbose_name=_("Resize"))
+    foar = models.CharField(verbose_name=_("Force Original Aspect Ratio"), max_length=10, choices=FOAR.choices(),
+                            default=FOAR.Disable)
+
+    queue = models.OneToOneField(Queue, null=True, blank=True, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True, verbose_name=_('Created Time'))
     updated_at = models.DateTimeField(auto_now=True, verbose_name=_('Last Update Time'))
@@ -84,7 +99,10 @@ class Schedule(models.Model):
     def __str__(self):
         return str(self.name)
 
-    def is_passed(self):
+    def get_foar(self) -> FOAR:
+        return FOAR.__members__.get(self.get_foar_display())
+
+    def is_passed(self) -> bool:
         return self.start_time <= timezone.now()
 
     def end_time(self):
@@ -95,142 +113,45 @@ class Schedule(models.Model):
         verbose_name = _("Schedule")
         verbose_name_plural = _("Schedules")
 
+    def save(self, **kwargs):
+        self.full_clean()
+        return super(Schedule, self).save(**kwargs)
 
-class TaskTypes(ChoiceEnum):
-    Record = 0
-    Resize = 1
-    AudioSync = 2
+    def delete(self, **kwargs):
+        if self.queue:
+            self.queue.delete()
+        return super(Schedule, self).delete(**kwargs)
 
+    def _set_status(self, stat: ScheduleStatus):
+        if self.status == stat:
+            logger.warning("Schedule<%d>: Status already %s can not change." % (self.id, stat.name))
+            return
 
-class Task(models.Model):
-    schedule = models.OneToOneField(Schedule)
-    type = models.SmallIntegerField(verbose_name=_("Type"), choices=[(t.value, _(t.name)) for t in TaskTypes])
-    completed = models.BooleanField(default=False)
-
-    created_at = models.DateTimeField(auto_now_add=True, verbose_name=_('Created Time'))
-    updated_at = models.DateTimeField(auto_now=True, verbose_name=_('Last Update Time'))
-
-    class Meta:
-        verbose_name = _("Task")
-        verbose_name_plural = _("Tasks")
-
-
-class BaseTaskModel(models.Model):
-    task = models.OneToOneField(Task)
-
-    log = models.TextField(verbose_name=_('Log'), null=True, blank=True)
-    pid = models.PositiveSmallIntegerField(null=True, blank=True)
-
-    DEFAULT_STATUS = None  # Overwrite This
-    STATUES = []  # Overwrite This
-    status = models.SmallIntegerField(verbose_name=_('Status'), choices=STATUES,
-                                      default=DEFAULT_STATUS)
-
-    started_at = models.DateTimeField(null=True, blank=True)
-    ended_at = models.DateTimeField(null=True, blank=True)
-
-    created_at = models.DateTimeField(auto_now_add=True, verbose_name=_('Created Time'))
-    updated_at = models.DateTimeField(auto_now=True, verbose_name=_('Last Update Time'))
-
-    class Meta:
-        abstract = True
-
-    @staticmethod
-    def _log_separator():
-        "-" * 10 + " " + timezone.now().strftime('%d/%m/%Y %H:%M:%S') + " " + "-" * 10 + "\n"
-
-    def add_log(self, msg: str, sep: bool = True, end: str = "\n", save: bool = True):
-        if msg:
-            log = self._log_separator() + str(msg) if sep else str(msg)
-            self.log = self.log + end + log if self.log else log
-            if save:
-                self.save(update_fields=['log'])
-        return self
-
-    def command(self):
-        """!Important: Overwrite this. This should generate command."""
-        raise NotImplemented()
-
-
-class RecordStatus(ChoiceEnum):
-    Canceled = -2
-    Error = -1
-    Waiting = 0
-    Recording = 1
-    Completed = 2
-
-
-class RecordTask(BaseTaskModel):
-    output = models.ForeignKey('Video', related_name="recordtask_output")
-
-    STATUES = [(s.value, _(s.name)) for s in RecordStatus]
-    DEFAULT_STATUS = RecordStatus.Waiting.value
-
-    class Meta:
-        verbose_name = _("Record")
-        verbose_name_plural = _("Records")
-
-    def command(self):
         try:
-            input = self.task.schedule.channel.url
+            logger.exception("Schedule<%d>: Changing status %s to %s" % (self.id, self.get_status_display(), stat.name))
+            self.status = stat.value
+            self.save(update_fields=['status'])
+        except Exception:
+            logger.exception("Schedule<%d>: can not change status to %s" % stat.name)
+            raise
 
-            output: Video = Video.objects.get_or_create(target_object_id=self.id,
-                                                        target_content_type=ContentType.objects.get_for_model(self))
-            output.create_file()
+    def set_status_scheduled(self):
+        self._set_status(ScheduleStatus.Scheduled)
 
-            cmd = Command({"input": input, "output": output.file.path, "loglevel": LogLevel.Error.value,
-                           "duration": self.task.schedule.time})
-            cmd.add_codec(Codec(data={"copy": True, "stream": StreamSpecifier.Video.value}))
+    def set_status_processing(self):
+        self._set_status(ScheduleStatus.Processing)
 
-            cmd.add_filter(
-                BitstreamChannelFilter({"stream": StreamSpecifier.Audio.value, "filters": FFmpegFilter.aac_adtstoasc})
-            )
-            cmd.validate()
-        except Exception as err:
-            logger.exception("Command can not created.")
-            raise err
+    def set_status_completed(self):
+        self._set_status(ScheduleStatus.Completed)
 
-        return cmd.generate()
+    def set_status_canceled(self):
+        self._set_status(ScheduleStatus.Canceled)
 
+    def set_status_timeout(self):
+        self._set_status(ScheduleStatus.TimeOut)
 
-class ResizeStatus(ChoiceEnum):
-    Canceled = -2
-    Error = -1
-    Started = 0
-    Ended = 1
-
-
-class ResizeTask(BaseTaskModel):
-    input = models.ForeignKey('Video', related_name="resizetask_input")
-    output = models.ForeignKey('Video', related_name="resizetask_output")
-
-    scale = models.CharField(max_length=10, choices=VIDEO_SIZES)
-    kar = models.BooleanField(default=False, verbose_name=_("Keep Aspect Ratio"))
-
-    STATUES = [(s.value, _(s.name)) for s in ResizeStatus]
-    DEFAULT_STATUS = ResizeStatus.Started.value
-
-    class Meta:
-        verbose_name = _("Resize")
-        verbose_name_plural = _("Resize")
-
-    def command(self):
-        """
-        :return: ffmpeg -i {input} -filter:v scale={scale} {output}
-        """
-        try:
-            output: Video = Video.objects.get_or_create(target_object_id=self.id,
-                                                        target_content_type=ContentType.objects.get_for_model(self))
-            output.create_file()
-            cmd = Command(data={"input": self.input.file.path, "output": output.file.path})
-            width, height = self.scale.split('x')
-            cmd.add_filter(
-                ScaleFilter(data={"width": int(width), "height": int(height), "stream": StreamSpecifier.Video.value}))
-            cmd.validate()
-        except Exception as err:
-            logger.exception("Command can not created.")
-            raise err
-        return cmd.generate()
+    def set_status_error(self):
+        self._set_status(ScheduleStatus.Error)
 
 
 class VideoFormat(ChoiceEnum):
@@ -241,23 +162,41 @@ class VideoFormat(ChoiceEnum):
 
 
 class Video(models.Model):
-    target_content_type = models.ForeignKey(ContentType, related_name='video_target', blank=True, null=True)
-    target_object_id = models.CharField(max_length=255, blank=True, null=True)
-    target = GenericForeignKey('target_content_type', 'target_object_id')
+    related_content_type = models.ForeignKey(ContentType, blank=True, null=True)
+    related_object_id = models.PositiveIntegerField(blank=True, null=True)
+    related = GenericForeignKey('related_content_type', 'related_object_id')
 
-    name = models.CharField(max_length=100, verbose_name=_("Video Name"))
-    file = models.FileField(verbose_name=_("Video File"), upload_to="videos/")
+    name = models.CharField(max_length=100, verbose_name=_("Video Name"), validators=[MinLengthValidator(1)])
+    file = models.FileField(verbose_name=_("Video File"), upload_to="videos/",
+                            validators=[FileExtensionValidator(VideoFormat.values())])
 
     format = models.CharField(max_length=5, verbose_name=_("File Format"), default=VideoFormat.MP4.value,
-                              choices=[(f.value, f.name) for f in VideoFormat])
+                              choices=VideoFormat.choices())
 
-    size = models.IntegerField(default=0, verbose_name=_("File Size"))
     attr = JSONField(verbose_name=_("Attributes"), null=True, blank=True)
-
-    is_empty = models.BooleanField(default=False)
 
     created_at = models.DateTimeField(auto_now_add=True, verbose_name=_("Create Time"))
     updated_at = models.DateTimeField(auto_now=True, verbose_name=_("Last Update"))
+
+    def size(self) -> int:
+        return int(self.file.size)
+
+    def set_related(self, obj):
+        try:
+            self.related_content_type = ContentType.objects.get_for_model(obj)
+            self.related_object_id = obj.pk
+            self.save()
+            return self
+        except Exception:
+            logger.exception("Video<%d>: Set target failed." % self.id)
+            raise
+
+    @staticmethod
+    def get_object_by_related(obj):
+        related_content_type = ContentType.objects.get_for_model(obj)
+        related_object_id = obj.pk
+        return Video.objects.all().filter(related_content_type=related_content_type,
+                                          related_object_id=related_object_id)
 
     def save_file_attributes(self):
         if not self.file:
@@ -266,21 +205,27 @@ class Video(models.Model):
         self.attr = get_file_attributes(self.file.path)
         self.save(update_fields=['attr'])
 
-    def create_file(self, ext: VideoFormat = None):
-        """Creates empty file as placeholder so can use `file.path` attribute"""
-        self.is_empty = True
-        self.size = 0
-        self.format = ext.value if ext else self.format
-        self.file.save(self.name + "." + self.format, ContentFile(''), save=False)
-        self.save(update_fields=['is_empty', 'file', 'size', 'format'])
+    def create_file(self):
+        try:
+            """Creates empty file as placeholder so can use `file.path` attribute"""
+            self.format = self.format
+            self.file.save("%s.%s" % (self.name, self.format), ContentFile(''), save=True)
+            return self
+        except Exception:
+            logger.exception("Video<%d>: Create file failed." % self.id)
+            raise
 
     def delete(self, **kwargs):
-        if self.file:
+        if self.file and os.path.exists(self.file.path):
             try:
                 os.remove(self.file.path)
             except Exception:
-                logger.exception("File could not deleted: %s" % self.file.path)
-        super(Video, self).delete(**kwargs)
+                logger.exception("Video<%d> File could not deleted: %s" % (self.id, self.file.path))
+        return super(Video, self).delete(**kwargs)
+
+    def save(self, **kwargs):
+        self.full_clean(exclude=['file'])
+        return super(Video, self).save(**kwargs)
 
     class Meta:
         verbose_name = _("Video")
